updating Swineherd::Script spec

+++ lib/swineherd/script.rb
@@ -1,8 +1,7 @@
-
 module Swineherd
   class Script
 
-    attr_reader   :raw_script_filename,:file
+    attr_reader   :raw_script_filename
     attr_accessor :binding
 
     def initialize(raw_script_filename,binding={})
@@ -11,7 +10,9 @@ module Swineherd
     end
 
     def run(settings={})
+      write
       runner.config.merge!(settings)
+      #Logger.new(STDOUT).info "\n#{evaluated_script_contents}"
       runner.execute
     end
 
@@ -20,28 +21,28 @@ module Swineherd
     end
 
     def flush!
-      File.delete(filename)
-      @file     = nil
-      @binding  = {}
-    end
-
-    def filename
-      write
-      file && file.path
+      File.delete(filename) if File.exists?(filename)
+      @template_filename = nil
+      @template_file = nil
+      @binding = {}
     end
 
     def write
-      @file ||= File.open(script_filename,"w+"){|file| file.write(script_contents);file}
+      template_file
+    end
+
+    def filename
+      @template_filename ||= Swineherd.config.template_root+[Time.now.to_i,$$,File.basename(raw_script_filename).gsub(/.erb$/,'')].join("-")
+    end
+
+    def evaluated_script_contents
+      Erubis::Eruby.new(File.read(raw_script_filename)).result(self.binding)
     end
 
     private
 
-    def script_filename
-      Swineherd.config.template_root+[Time.now.to_i,$$,File.basename(raw_script_filename).gsub(/.erb$/,'')].join("-")
-    end
-
-    def script_contents
-      Erubis::Eruby.new(File.read(raw_script_filename)).result(self.binding)
+    def template_file
+      @template_file ||= File.open(filename,"w+"){|file| file.write(evaluated_script_contents);file}
     end
 
   end
